libavcodec/qsvenc: Add support to qsv to encode external surface.
Qsv encoder only encode the frame that are pre-allocated, so qsv encoder
cannot encode the frame mapped from vulkan. In fact, MediaSDK can encode
frame that are dynamically created. I add the support for qsv to encode
external frame. The resource will be allocated and released in real time
if the encoded frame is not in the frame pool.

Signed-off-by: Zhizhen Tang <zhizhen.tang@intel.com>
Signed-off-by: Wenbin,Chen <wenbin.chen@intel.com>
Signed-off-by: Tong Wu <tong1.wu@intel.com>